Include MEPA code to type of the passage
Also include code to return of the function.
Now need to stack up vars with correct "nivel lexico"
and "deslocamento" of the passage.
Need to remove locals vars, paremeters and procedure
or function when out of procedures or functions, to vars
and paremeters need to remove when out of procedure or
function, to procedures and fuctions need remove when
out to level ups of they.
Need to implement read and generate MEPA code to "goto".

Signed-off-by: Gustavo Toshi Komura <gtk10@c3sl.ufpr.br>